+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## [2.2.0] 2020-11-02
+
+### Added
+
+- create tables
+- create service definitions
+
 ## [2.1.1] 2020-10-23
 
 ### Fixed